Move cli utils in the aiida_quantumespresso module
This allows other packages that have this package as a
dependency to also use the cli utilities to write their own
command line interface
